Add BeforeAndAfterEach
