+++ resource-managers/kubernetes/integration-tests/src/test/scala/org/apache/spark/deploy/k8s/integrationtest/KubernetesSuite.scala
@@ -197,20 +197,16 @@ class KubernetesSuite extends SparkFunSuite
     }
   }
 
-  protected def cleanUpTest(): Unit = {
-    if (!kubernetesTestComponents.hasUserSpecifiedNamespace) {
-      kubernetesTestComponents.deleteNamespace()
-    }
-    deleteDriverPod()
-    deleteExecutorPod()
-  }
-
   before {
     setUpTest()
   }
 
   after {
-    cleanUpTest()
+    if (!kubernetesTestComponents.hasUserSpecifiedNamespace) {
+      kubernetesTestComponents.deleteNamespace()
+    }
+    deleteDriverPod()
+    deleteExecutorPod()
   }
 
   protected def runSparkPiAndVerifyCompletion(
